samples/python/module_genai/requirements.txt
pyyaml
graphviz
numpy
pillow
torch
torchvision
transformers
git+https://github.com/huggingface/diffusers
git+https://github.com/openvino-dev-samples/optimum-intel.git@zimage